Split the code so that it isn't in one index.js file
