+++ .dockerignore
@@ -8,4 +8,3 @@ LICENSE
 .env.example
 .github
 CONTRIBUTING.md
-renovate.json
